LF-4790 Tweaks * rename IPDetailKPIProps to IrrigationPrescriptionKPI
